refactor/filter: Move count_n_unique into utils.search_utils.stats
